+++ src/components/ErrorMessage/index.ts
@@ -1,2 +1,2 @@
 export { default as ErrorMessage } from './ErrorMessage.vue'
-export type { ErrorMessageProps } from './types'
+export type { ErrorMessageProps, ErrorMessageValue } from './types'